My original push to github apparently failed. When I try to push now I get strange errors. Its like github knows of my commits yet it's not showing on the website nor does it come down when you clone or pull.
I made a patch against the version in github and my local code based, applied
it in one shot on a clean clone and now hoping this brand new commit will push
correctly (all the history of this new work is lost now but at least the code
is there...)

This man have been caused by the extremely poor wifi at the airport...

This code is all the changes from winter break mainly a few small bug fixes and
some small code scrubbing. Also the overlay class no longer holds its own copy
of the rock object. It just knows of the selected rock id and it gets a new one
each time.

The initial commit was 12/16/2012 just in case the original show up sometime...